[data-refactor] docs : repository, usecase domain 패키지로 이동
