+++ app/src/main/java/com/hegunhee/subwayarrivalinfoapp/domain/usecase/DeleteFavoritesUseCase.kt
@@ -1,9 +1,6 @@
-package com.hegunhee.subwayarrivalinfoapp.domain
+package com.hegunhee.subwayarrivalinfoapp.domain.usecase
 
-import com.hegunhee.subwayarrivalinfoapp.model.Repository
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
+import com.hegunhee.subwayarrivalinfoapp.domain.repository.Repository
 import javax.inject.Inject
 
 class DeleteFavoritesUseCase @Inject constructor(private val repository: Repository){
