+++ app/src/main/java/com/hegunhee/subwayarrivalinfoapp/domain/usecase/GetFavoriteSubwayInfoUseCase.kt
@@ -1,8 +1,8 @@
-package com.hegunhee.subwayarrivalinfoapp.domain
+package com.hegunhee.subwayarrivalinfoapp.domain.usecase
 
 import com.hegunhee.subwayarrivalinfoapp.data.entity.Favorites
 import com.hegunhee.subwayarrivalinfoapp.data.json.subway_arrival.SubwayArrivalSmallDataWithFavorite
-import com.hegunhee.subwayarrivalinfoapp.model.Repository
+import com.hegunhee.subwayarrivalinfoapp.domain.repository.Repository
 import javax.inject.Inject
 
 class GetFavoriteSubwayInfoUseCase @Inject constructor(private val repository : Repository) {
