+++ app/src/main/java/com/hegunhee/subwayarrivalinfoapp/domain/usecase/SaveAllSubwayListInLocalDBUseCase.kt
@@ -1,6 +1,6 @@
-package com.hegunhee.subwayarrivalinfoapp.domain
+package com.hegunhee.subwayarrivalinfoapp.domain.usecase
 
-import com.hegunhee.subwayarrivalinfoapp.model.Repository
+import com.hegunhee.subwayarrivalinfoapp.domain.repository.Repository
 import javax.inject.Inject
 
 class SaveAllSubwayListInLocalDBUseCase @Inject constructor(private val repository: Repository){
