+++ app/src/main/java/com/hegunhee/subwayarrivalinfoapp/ui/favorite/FavoriteViewModel.kt
@@ -3,8 +3,8 @@ package com.hegunhee.subwayarrivalinfoapp.ui.favorite
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.hegunhee.subwayarrivalinfoapp.data.entity.Favorites
-import com.hegunhee.subwayarrivalinfoapp.domain.DeleteFavoritesUseCase
-import com.hegunhee.subwayarrivalinfoapp.domain.GetFavoritesListByFlowUseCase
+import com.hegunhee.subwayarrivalinfoapp.domain.usecase.DeleteFavoritesUseCase
+import com.hegunhee.subwayarrivalinfoapp.domain.usecase.GetFavoritesListByFlowUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
